+++ lib/extwitter/config.ex
@@ -29,7 +29,7 @@ defmodule ExTwitter.Config do
   def get_tuples do
     case ExTwitter.Config.get do
       nil -> []
-      tuples -> tuples |> Enum.map(fn({k, v}) -> {k, to_char_list(v)} end)
+      tuples -> tuples
     end
   end
 end