+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 0.8.0
 ------
 #### Changes (Breaking)
-* API Change - Major changes related to ExTwitter raising errors.
+* API Change - Major changes related to ExTwitter raising errors (#56).
     - Most calls will now return an {:error, message} tuple
-      instead of raising an error directly
+      instead of raising an error directly.
+
+* OAuth dependency library is updated (#63).
+    - Adding `{:oauth, github: "tim/erlang-oauth"}` in mix.exs becomes not required.
+
+* Configurations for proxy is simplified.
+    - https://github.com/parroty/extwitter#proxy
 
 0.7.3
 ------